% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_impulses.R
\name{compute_impulses}
\alias{compute_impulses}
\title{Compute impulse responses}
\usage{
compute_impulses(Bh, swish, nn)
}
\arguments{
\item{Bh}{A draw of the parameter matrix}

\item{swish}{A draw of the covariance matrix}

\item{nn}{A vector containging the amount of variables in the sytem, number
of lags and prediction horizon.}
}
\value{
A matrix with impulse responses
}
\description{
Taking the draws of parameter matrix and covariance matrix and compute impulse
responses for a predefined number of steps.
}
