% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebr_accept.R
\name{ebr_accept}
\alias{ebr_accept}
\title{Check whether elasticity bounds hold}
\usage{
ebr_accept(EBR = NULL, impulses = NULL, Q = NULL, allshocknames)
}
\arguments{
\item{EBR}{Object of ebt_setup}

\item{impulses}{Computed impulses based on parameter draws}

\item{Q}{A orthogonal matrix}
}
\value{
A Boolean whether elasticity bounds are fulfilled
}
\description{
Check whether elasticity bounds hold
}
\examples{

}
