% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebr_setup.R
\name{ebr_setup}
\alias{ebr_setup}
\title{Setup the elasticitiy bound restrictions}
\usage{
ebr_setup(
  shock_names = NULL,
  first_variable = NULL,
  second_variable = NULL,
  horizon = NULL,
  bounds = NULL
)
}
\arguments{
\item{shock_names}{A vector of shock names on which elasticity bounds are
placed.}

\item{first_variable}{A vector of variable names of length shock names}

\item{second_variable}{A vector of variables names of length shock names}

\item{horizon}{A vector of horizons for which elasticity bound restrictions
should hold}

\item{bounds}{A vectir of numerics which specifiy the elasticity bounds}
}
\description{
This function sets up the elasticity bound restrictions.
}
