% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fevd_plot.R
\name{fevd_plot}
\alias{fevd_plot}
\title{Plot the forecast error variance decompositions}
\usage{
fevd_plot(
  narrSign_model = NULL,
  whichShock = NULL,
  narr = FALSE,
  type = "median",
  bands = c(0.16, 0.84),
  varnames = NULL
)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}

\item{whichShock}{A string with the name of shock to be plotted}

\item{narr}{A Boolean whether shocks identified by narrative restrictions
should be used. If False only traditional}
}
\value{
A ggplot object which displays the FEVD for a specific shock.
}
\description{
Plot the forecast error variance decompositions
}
