% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_plot.R
\name{HD_plot}
\alias{HD_plot}
\title{Plot historical decomposition for a specific date}
\usage{
HD_plot(narrSign_model = NULL, whichShock = NULL, date = NULL)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}

\item{whichShock}{A string with the name of shock to be plotted}

\item{date}{A string indicating which historical decompositions to plot}
}
\value{
A ggplot object
}
\description{
Plot historical decomposition for a specific date
}
\examples{
\dontrun{
HD_plot(model, whichShock = 1, data = "2020-04-01")
}
}
