% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_plot.R
\name{irf_plot}
\alias{irf_plot}
\title{Plot IRFs}
\usage{
irf_plot(
  narrSign_model = NULL,
  whichShock = NULL,
  cumulative = NULL,
  narr = FALSE,
  type = "median",
  bands = c(0.16, 0.84),
  scaling = 1,
  varnames = NULL,
  steps = dim(narrSign_model$trad$IRFS)[2]
)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}

\item{whichShock}{A string with the name of shock to be plotted or a numeric indicating the number}

\item{cumulative}{A Boolean whether IRFs should be cumulative or not}

\item{narr}{A Boolean whether IRFs should be plotted which fulfill narrative or
only traditional restrictions}
}
\value{
A ggplot object
}
\description{
Plots all IRFs for a specified shock
}
