% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kilian_2009}
\alias{kilian_2009}
\title{Kilian 2009 replication data}
\format{
A tibble with 515 Rows and 7 variables:

\describe{
\item{dates}{Monthly dates in yyyy-mm-dd format}
\item{oil_production_growth}{Growth of oil worldwide oil production}
\item{economic_activity}{Economic activity index by Lutz Kilian}
\item{cprindex}{Real price of oil measured as the import price of crude
oil}
}
}
\source{
replication data Diaz Ramirez et al. (2018)
}
\usage{
kilian_2009
}
\description{
A dataset containing one column of dates and 6 variables.
}
\references{
Kilian 2009
}
\keyword{datasets}
