% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrsign_accept.R
\name{narrsign_accept}
\alias{narrsign_accept}
\title{Narrative restrictions accept and reweight}
\usage{
narrsign_accept(narrative_restr, irfs, shocks, ndrawweights, data, lag, fevds)
}
\arguments{
\item{narrative_restr}{Narrative restriction as output from narrsign_setup}

\item{irfs}{A 4 dimensional array which contains IRFs for all sstructural shocks}

\item{shocks}{A 3 dimensional array of structural shocks}

\item{ndrawweights}{A numeric with the amount of draws for the fake shocks}

\item{data}{Data used to estimate the model}

\item{lag}{A numeric with lag used in the system}

\item{fevds}{}
}
\description{
Checks narrative restrictions and reweights them accordingly.
}
\details{
Will in the long run change this such that this all comes from the model
}
