% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrsign_print.R
\name{narrsign_print}
\alias{narrsign_print}
\title{Print model for SVAR model identified with sign}
\usage{
narrsign_print(narrSign_model = NULL)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}
}
\value{
some model information printed to console
}
\description{
Print model for SVAR model identified with sign
}
