% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrsign_setup.R
\name{narrsign_setup}
\alias{narrsign_setup}
\title{Setup narrative sign restrictions}
\usage{
narrsign_setup(
  allshocknames = NULL,
  shock_names = NULL,
  shock_type = NULL,
  shock_dates = NULL,
  dates = NULL,
  relevant_variable = NULL,
  shock_sign = NULL,
  shock_size = NULL,
  lag = NULL,
  data = NULL
)
}
\arguments{
\item{allshocknames}{A vector containing all named shocks in the system}

\item{shock_names}{A vector of shock names on which narrative restrictions
are placed}

\item{shock_type}{A vector of type of shocks which corresponds to the name
of shocks}

\item{shock_dates}{A matrix of dimensions 2 x (length of shock_names) which
contains start and end date of narrative restriction}

\item{dates}{A vector of dates based on the original data set}

\item{relevant_variable}{An integer or string
the name of variable on which the narrative restriction is placed}

\item{shock_sign}{Either 1 or -1 depdning on the sign of the shock}

\item{shock_size}{A string which further specifies the kind of shock}

\item{lag}{A numeric for the amount of lags in the system}

\item{data}{Data used for estimation. Note: Without the dates!}
}
\value{
A list of arrays which contains the narrative restrictions
}
\description{
This functions takes the user input about narrative sign restrictions and
sets up the arrays necessary for evaluation. Two types of restrictions are
possible and should be noted carefully. One can either place a restriction
on the sign of a specific shock or on the historic contribution. See package
documention and the examples which should be self explanatory.
}
