% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducedform_var.R
\name{reducedform_var}
\alias{reducedform_var}
\title{Estimate a reduced form VAR model}
\usage{
reducedform_var(
  ydata = NA,
  lags = 2,
  xdata = NULL,
  const = TRUE,
  breaks = NULL
)
}
\arguments{
\item{ydata}{A Matrix containing data}

\item{lags}{Numeric of amount of lags to include in the model}

\item{xdata}{External data. (Not supported in narrSign package)}

\item{const}{A Boolean whether to include a constant in the model}

\item{breaks}{Breaks (Not supoerted in narrSign package)}
}
\value{
A list containing paremeter, covariance, residuals, singularity check
}
\description{
This function estimates the reduced form VAR model.
}
