% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock_plot.R
\name{shock_plot}
\alias{shock_plot}
\title{Plot structural shocks for all identified shocks}
\usage{
shock_plot(narrSign_model = NULL, whichShock = NULL, narr = FALSE, prob = 0.5)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}

\item{whichShock}{A string with the name of shock to be plotted}

\item{narr}{A Boolean whether shocks identified by narrative restrictions
should be used. If False only traditional}
}
\value{
A ggplot object
}
\description{
Plot structural shocks for all identified shocks
}
