% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradsign_accept.R
\name{tradsign_accept}
\alias{tradsign_accept}
\title{Check whether traditional sign restrictions hold}
\usage{
tradsign_accept(first, last, constrained, impulses, Q, signshocks, cum)
}
\arguments{
\item{first}{A numeric which is the first period in which the restrictions
should hold}

\item{last}{A numeric which is the last period in which the restrictions
should hold}

\item{constrained}{A K X 1 vector containing the traditional restrictions}

\item{impulses}{A matrix containging the computed impulse responses}

\item{Q}{An orthogonal matrix}

\item{signshocks}{Which column of the Q matrix is considered}
}
\value{
A list numeric (1 if accepted, 0 if not accpeted)
}
\description{
Takes impulse responses and Q draw to check whether traditional sign restrictions
are fulfilled
}
