% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradsign_estim.R
\name{tradsign_estim}
\alias{tradsign_estim}
\title{Title}
\usage{
tradsign_estim(
  Y = NULL,
  nlags = 4,
  draws = 200,
  subdraws = 200,
  nkeep = 1000,
  tradsign_setup = NULL,
  constant = TRUE,
  steps = 24,
  EBR = NULL,
  oil_production = NULL
)
}
\arguments{
\item{Y}{Data for estimation without dates}

\item{nlags}{A numeric for the amount of lags}

\item{draws}{A numeric for amount of draws}

\item{subdraws}{A numeric for amount of subdrwas}

\item{nkeep}{A numeric for amount of traditionally identified to keep}

\item{tradsign_setup}{A object of class tradsign_setup}

\item{constant}{A Boolean on whether constant should be included}

\item{steps}{A numeric for amount of steps}

\item{EBR}{An object of function ebr_setup}
}
\description{
Title
}
