% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradsign_setup.R
\name{tradsign_setup}
\alias{tradsign_setup}
\title{Setup up of traditional sign restrictions.}
\usage{
tradsign_setup(
  data = NULL,
  shock_names = NULL,
  restr_matrix = NULL,
  hor_matrix = NULL,
  cum = NULL
)
}
\arguments{
\item{data}{Orignal data used for the model.}

\item{shock_names}{A vector of shock names on which traditional sign
restrictions. Notice this can be greather than the number of variables in
case one wants to place different kinds of restrictions with different horizons
on each structural shock.}

\item{restr_matrix}{A matrix of dimensions shock names x K where each row
represent one sign restriction.}

\item{hor_matrix}{A matrix which defines the hoirzons for which the restrictions
are supposed to hold}
}
\value{
A list which contains restrictions hoirzons and shock names
}
\description{
This function sets up the traditional sign restrictions. Notice that
these restrictions can even be more than 1 for each structural shock.
}
