% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_fevd_plot.R
\name{two_fevd_plot}
\alias{two_fevd_plot}
\title{Plot the forecast error variance decompositions for both traditional and narrative
restrictions}
\usage{
two_fevd_plot(
  narrSign_model = NULL,
  whichShock = NULL,
  varnames = NULL,
  type = "median",
  bands = c(0.16, 0.84),
  steps = dim(narrSign_model$trad$IRFS)[2]
)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}

\item{whichShock}{A string with the name of shock to be plotted}
}
\description{
Plot the forecast error variance decompositions for both traditional and narrative
restrictions
}
