% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_irf_plot.R
\name{two_irf_plot}
\alias{two_irf_plot}
\title{Plot IRFs for traditional and narrative restrictions}
\usage{
two_irf_plot(
  narrSign_model = NULL,
  whichShock = NULL,
  cumulative = NULL,
  scaling_trad = 1,
  scaling_nar = 1,
  varnames = NULL,
  type = "median",
  bands = c(0.16, 0.84),
  steps = dim(narrSign_model$trad$IRFS)[2]
)
}
\arguments{
\item{narrSign_model}{A model of class narrsign}

\item{whichShock}{A numeric indicating which shock to plot. Based on the order in
narrSign_model}

\item{cumulative}{A Boolean whether cumulative IRFs should be plotted}
}
\value{
A ggplot object
}
\description{
Plot IRFs for traditional and narrative restrictions
}
