% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlig}
\alias{uhlig}
\title{Uhlig replication data}
\format{
A tibble with X Rows and 7 variables;
}
\source{
replication data Diaz Ramirez et al. (2018)
}
\usage{
uhlig
}
\description{
A dataset containing one column of dates and 6 variables.
}
\keyword{datasets}
