% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrsign.R
\name{narrsign}
\alias{narrsign}
\title{Estimate a SVAR model with narrative restrictions}
\usage{
narrsign(
  data = NULL,
  lags = 12,
  trad_signs = NULL,
  narr_restr = NULL,
  nkeep = 1000,
  draws = 300,
  subdraws = 300,
  elasticitybounds = NULL,
  const = TRUE,
  steps = 30,
  narrweightdraws = 1000
)
}
\arguments{
\item{data}{A tibble or ts object containing the data}

\item{lags}{A numeric with the amound of lags to include}

\item{trad_signs}{An object as output from tradsign_setup}

\item{narr_restr}{An object as output from narrsign_setup. NULL if no narrative
restrictions should be considered}

\item{nkeep}{Amount of draws to keep for the traditional sign restrictions}

\item{draws}{Amount of draws to make for parameter and covariance matrix}

\item{subdraws}{Amount of subdraws for orthogonal matrix Q for each
parameter matrix and covariance matrix}

\item{elasticitybounds}{An object as output from ebr_setup. NULL if no elasticity
bound should be considered}

\item{const}{Boolean whether constant should be considered}

\item{steps}{Amount of steps for which to compute the impulse responses}

\item{narrweightdraws}{Amount of draws for the importance sampler}
}
\value{
A list with outputs differentiating between model identified with
traditional sign restrictions and narrative sign restictions.
}
\description{
Estimate a SVAR model with narrative restrictions
}
